🎨 Refactor heartbeat into separate file HeartbeatHandler.
