--- src/main/java/org/websocket/controllers/PVcontroller.java
@@ -1,9 +1,0 @@
-package org.websocket.controllers;
-
-public class PVcontroller {
-
-
-
-
-
-}
