+++ src/main/java/org/websocket/handlers/SubscriptionHandler.java
@@ -1,4 +1,4 @@
-package org.websocket;
+package org.websocket.handlers;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
