+++ src/main/java/org/websocket/util/MetaDataCache.java
@@ -1,16 +1,12 @@
 package org.websocket.util;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import org.websocket.SessionHandler;
+import org.websocket.handlers.SessionHandler;
 import org.websocket.models.PV;
 import org.websocket.models.PvMetaData;
 
-import java.lang.reflect.Field;
 import java.util.HashMap;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.Executors;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.TimeUnit;
 
 
 // HAVE OBJECT CALLED PV META DATA, AND MAP IT WITH THE JSON PARSER
